requirements.txt
jira==1.0.10
requests>=2.10.0
pyyaml>=3.12